Updates to analysis + upload of presentation
Updated the model in the analysis R script: There was a flaw in the model code, meaning that the wrong column was used for the outcome variable. This has been changed, resulting in an entirely new model, with a way less uncertain prediction and a prediction in the hypothesised direction.

The presentation for the oral part of the exam has been uploaded to the repository.